[feat] 외부 클릭시 닫히는 커스텀 훅
