+++ src/pages/UserInfo/LoginPage.jsx
@@ -1,16 +1,13 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import { useNavigate } from "react-router-dom";
 
-import IntroTopbar from "../components/IntroTopbar";
+import IntroTopbar from "../../components/IntroTopbar";
 
-import { ReactComponent as Symbol } from "../images/symbol.svg";
-import { ReactComponent as Watermark } from "../images/watermark-kor.svg";
-import { ReactComponent as WatermarkE } from "../images/watermark-eng.svg";
-import { ReactComponent as Kakao } from "../images/kakao.svg";
+import { ReactComponent as Symbol } from "../../images/symbol.svg";
+import { ReactComponent as Watermark } from "../../images/watermark-kor.svg";
+import { ReactComponent as WatermarkE } from "../../images/watermark-eng.svg";
 
-const InitialPage = () => {
-  const navigate = useNavigate();
+const LoginPage = () => {
   return (
     <Wrapper>
       <IntroTopbar />
@@ -27,32 +24,17 @@ const InitialPage = () => {
       </Box>
       <SideBox>
         <Login>로그인</Login>
-        <BtnBox>
-          <SignupBtn
-            onClick={() => {
-              navigate("/signup");
-            }}
-          >
-            회원가입
-          </SignupBtn>
-          <LoginBtn
-            onClick={() => {
-              navigate("/login");
-            }}
-          >
-            로그인
-          </LoginBtn>
-          <KakaoBtn>
-            <Kakao />
-            카카오 로그인
-          </KakaoBtn>
-        </BtnBox>
+        <InputBox>
+          <Input type="text" placeholder="아이디" />
+          <Input type="password" placeholder="비밀번호" />
+        </InputBox>
+        <LoginBtn>로그인</LoginBtn>
       </SideBox>
     </Wrapper>
   );
 };
 
-export default InitialPage;
+export default LoginPage;
 
 const Wrapper = styled.div`
   display: flex;
@@ -118,8 +100,8 @@ const SideBox = styled(ColumnStyle)`
     padding: 10rem;
     flex-direction: column;
     align-items: flex-start;
-    border-radius: 1.6rem;
-    border: 1.5px solid var(--Light-Gray, #fafafa);
+    border-radius: 16px;
+    border: 1.5px solid var(--lightGray);
     background: var(--white);
 
     box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.1);
@@ -127,7 +109,7 @@ const SideBox = styled(ColumnStyle)`
 `;
 
 const Login = styled.div`
-  color: var(--Black, #262121);
+  color: var(--black);
   text-align: center;
 
   font-size: 3.2rem;
@@ -140,41 +122,49 @@ const Login = styled.div`
   }
 `;
 
-const BtnBox = styled(ColumnStyle)`
-  gap: 1rem;
+const InputBox = styled(ColumnStyle)`
   width: 100%;
+  gap: 1.6rem;
 `;
 
-const Btn = styled.div`
-  display: flex;
-  width: 29.8rem;
-  padding: 2rem 4.8rem;
-  justify-content: center;
-  align-items: center;
+const Input = styled.input`
+  width: 100%;
+  height: 4.8rem;
+  background: var(--white);
+  padding: 0 1rem;
 
-  border-radius: 1.2rem;
+  font-size: 1.6rem;
+  font-style: normal;
+  font-weight: 500;
 
-  font-size: 20px;
-  font-weight: 600;
-  cursor: pointer;
+  border: none;
+  outline: none;
+  border-bottom: 0.2rem solid var(--black);
+  transition: border-bottom-color 0.3s ease;
 
-  @media (min-width: 1200px) {
-    width: 100%;
+  &::placeholder {
+    color: var(--gray);
   }
 `;
 
-const SignupBtn = styled(Btn)`
-  background: var(--Black, #262121);
-  color: var(--White, #fff);
-`;
-
-const LoginBtn = styled(Btn)`
-  border: 1.5px solid var(--Gray, #d9d9d9);
-  background: var(--White, #fff);
-`;
-
-const KakaoBtn = styled(Btn)`
-  background: #fee500;
-  color: #3b1d1d;
+const LoginBtn = styled.button`
+  display: flex;
+  padding: 2rem 4.8rem;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
   gap: 1.6rem;
+  align-self: stretch;
+
+  border-radius: 1.2rem;
+  border: 0.15rem solid var(--gray);
+  background: var(--white);
+
+  margin: 3.2rem 0;
+
+  color: var(--black);
+  text-align: center;
+
+  font-size: 2rem;
+  font-weight: 600;
 `;
